shanoir-issue#2188: call async method from another class
